nixos-iso: Add Hyper-V Support
Version 2 Hyper-V containers require that paravirtualize scsi driver be
used. This driver is specific to Hyper-V and needed to detect the CD-ROM
for the iso during the Stage-1.

Additionally, Hyper-V requires a keyboard driver in order to interact
with the TTY upon booting the ISO. We add this to the initrd in order to
provide keyboard access in case the Stage-1 fails.